image-generator/requirements.txt
flask==3.0.0
flask-cors==4.0.0
pillow==10.2.0
torch==2.0.1
transformers==4.36.2
diffusers==0.25.0
accelerate==0.25.0
safetensors==0.4.1